fix: WA Security MCP server architecture review (13 items)
Critical:
- Fix ARN validation to support aws-us-gov and aws-cn partitions
- Document context_storage as ephemeral (lost on restart)

High:
- Add pyproject.toml with pinned deps and entry point
- Cache AWS session with 50-min TTL in create_aws_session()
- Remove access analyzer override hack (underlying fn already correct)
- Split server.py from 1251 to 104 lines into tools/ and prompts.py

Medium:
- Delete dead code: environment_models.py, prompt_utils.py
- Fix util/__init__.py exports (add macie, trusted advisor, etc.)
- Replace all print(DEBUG) with loguru logger calls
- Simplify check_inspector() from 242 to 86 lines (use batch_get_account_status)

Low:
- CloudFront region gating returns message instead of silent skip
- Host configurable via MCP_HOST env var and --host CLI arg
- generate_recommendations() references specific resource counts

All 234 tests pass.
