+++ mcp-servers/well-architected-security-mcp-server-with-iamrole-support/pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "well-architected-security-mcp-server"
+version = "0.1.0"
+requires-python = ">=3.10"
+license = "MIT-0"
+dependencies = [
+    "boto3==1.38.18",
+    "botocore==1.38.18",
+    "loguru==0.7.3",
+    "mcp==1.11.0",
+    "pydantic==2.10.6",
+    "fastmcp==2.12.4",
+]
+
+[project.scripts]
+wa-security-mcp-server = "src.server:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src"]
