--- mcp-servers/well-architected-security-mcp-server-with-iamrole-support/tests/test_prompt_utils.py
@@ -1,419 +1,0 @@
-# MIT No Attribution
-#
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of
-# this software and associated documentation files (the "Software"), to deal in
-# the Software without restriction, including without limitation the rights to
-# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-# the Software, and to permit persons to whom the Software is furnished to do so.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Tests for the prompt_utils module."""
-
-import os
-import tempfile
-from unittest import mock
-
-import pytest
-
-from src.util.prompt_utils import (
-    get_all_template_names,
-    get_prompt_template,
-    get_template_metadata,
-    load_prompt_templates,
-)
-
-
-@pytest.fixture
-def reset_prompt_utils_state():
-    """Reset the global state in prompt_utils module between tests."""
-    # Import the module to access its globals
-    from src.util import prompt_utils
-
-    # Save original values
-    original_templates = prompt_utils._prompt_templates
-    original_initialized = prompt_utils._is_initialized
-
-    # Reset to initial state
-    prompt_utils._prompt_templates = {}
-    prompt_utils._is_initialized = False
-
-    yield
-
-    # Restore original values
-    prompt_utils._prompt_templates = original_templates
-    prompt_utils._is_initialized = original_initialized
-
-
-def test_load_prompt_templates_file_not_found(reset_prompt_utils_state):
-    """Test loading templates when file doesn't exist."""
-    # Test with non-existent file
-    templates = load_prompt_templates("nonexistent_file.md")
-    assert templates == {}
-
-
-def test_load_prompt_templates_valid_file(reset_prompt_utils_state):
-    """Test loading templates from a valid markdown file."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-with multiple lines
-```
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-
-## Example Workflow
-This is an example workflow section
-
-Step 1: Do something
-Step 2: Do something else
-""")
-        temp_path = temp.name
-
-    try:
-        # Load templates from the temp file
-        templates = load_prompt_templates(temp_path)
-
-        # Assert correct parsing
-        assert len(templates) == 4  # Actual implementation returns 4 templates
-        assert "template_1" in templates
-        assert "template_2" in templates
-        assert "workflow_example" in templates
-
-        # Check template 1
-        assert templates["template_1"]["content"] == "Template 1 content\nwith multiple lines"
-        assert templates["template_1"]["description"] == "Description for template 1"
-        assert templates["template_1"]["title"] == "Template 1"
-
-        # Check template 2
-        assert templates["template_2"]["content"] == "Template 2 content"
-        assert templates["template_2"]["description"] == "Description for template 2"
-
-        # Check workflow example
-        assert "Step 1: Do something" in templates["workflow_example"]["content"]
-        assert templates["workflow_example"]["title"] == "Example Workflow"
-    finally:
-        # Clean up the temp file
-        os.unlink(temp_path)
-
-
-def test_load_prompt_templates_empty_file(reset_prompt_utils_state):
-    """Test loading templates from an empty file."""
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp_path = temp.name
-
-    try:
-        templates = load_prompt_templates(temp_path)
-        assert templates == {}
-    finally:
-        os.unlink(temp_path)
-
-
-def test_load_prompt_templates_invalid_format(reset_prompt_utils_state):
-    """Test loading templates from a file with invalid format."""
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-This is not a valid template format.
-No section headers or code blocks.
-""")
-        temp_path = temp.name
-
-    try:
-        templates = load_prompt_templates(temp_path)
-        assert templates == {}
-    finally:
-        os.unlink(temp_path)
-
-
-def test_load_prompt_templates_missing_code_block(reset_prompt_utils_state):
-    """Test loading templates with missing code blocks."""
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        templates = load_prompt_templates(temp_path)
-        assert len(templates) == 1  # Only template_2 should be loaded
-        assert "template_2" in templates
-    finally:
-        os.unlink(temp_path)
-
-
-def test_load_prompt_templates_caching(reset_prompt_utils_state):
-    """Test that templates are cached after first load."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        # First load
-        templates1 = load_prompt_templates(temp_path)
-        assert "template_1" in templates1
-
-        # Modify the file
-        with open(temp_path, "w") as f:
-            f.write("""# Test Templates
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-""")
-
-        # Reset initialization flag and load again
-        from src.util import prompt_utils
-
-        prompt_utils._is_initialized = False
-        templates3 = load_prompt_templates(temp_path)
-        assert "template_2" in templates3
-    finally:
-        os.unlink(temp_path)
-
-
-def test_get_prompt_template(reset_prompt_utils_state):
-    """Test getting a specific prompt template."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-```
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        # Load templates first
-        load_prompt_templates(temp_path)
-
-        # Get specific template
-        template = get_prompt_template("template_1")
-        assert template is not None
-        assert template["content"] == "Template 1 content"
-
-        # Get non-existent template
-        template = get_prompt_template("nonexistent_template")
-        assert template is None
-    finally:
-        os.unlink(temp_path)
-
-
-def test_get_all_template_names(reset_prompt_utils_state):
-    """Test getting all template names."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-```
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        # Load templates first
-        load_prompt_templates(temp_path)
-
-        # Get all template names
-        names = get_all_template_names()
-        assert len(names) == 2
-        assert "template_1" in names
-        assert "template_2" in names
-    finally:
-        os.unlink(temp_path)
-
-
-def test_get_template_metadata(reset_prompt_utils_state):
-    """Test getting metadata for all templates."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-```
-
-## Template 2
-Description for template 2
-
-```
-Template 2 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        # Load templates first
-        load_prompt_templates(temp_path)
-
-        # Get template metadata
-        metadata = get_template_metadata()
-        assert len(metadata) == 2
-
-        # Check first template metadata
-        template1_meta = next(m for m in metadata if m["name"] == "template_1")
-        assert template1_meta["title"] == "Template 1"
-        assert template1_meta["description"] == "Description for template 1"
-
-        # Check second template metadata
-        template2_meta = next(m for m in metadata if m["name"] == "template_2")
-        assert template2_meta["title"] == "Template 2"
-        assert template2_meta["description"] == "Description for template 2"
-    finally:
-        os.unlink(temp_path)
-
-
-def test_load_prompt_templates_with_exception(reset_prompt_utils_state):
-    """Test handling of exceptions during template loading."""
-    with mock.patch("os.path.exists", return_value=True):
-        with mock.patch("builtins.open", side_effect=Exception("Test exception")):
-            templates = load_prompt_templates("test.md")
-            assert templates == {}
-
-
-def test_get_prompt_template_loads_if_not_initialized(reset_prompt_utils_state):
-    """Test that get_prompt_template loads templates if not already initialized."""
-    # Create a temp file with test content
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".md") as temp:
-        temp.write("""# Test Templates
-
-## Template 1
-Description for template 1
-
-```
-Template 1 content
-```
-""")
-        temp_path = temp.name
-
-    try:
-        # Mock the default file path
-        with mock.patch("src.util.prompt_utils.load_prompt_templates") as mock_load:
-            mock_load.return_value = {"template_1": {"content": "mocked content"}}
-
-            # Call get_prompt_template without loading templates first
-            template = get_prompt_template("template_1")
-
-            # Verify load_prompt_templates was called
-            mock_load.assert_called_once()
-            # The actual implementation might return None if the template doesn't exist
-            # Let's adjust our assertion to match the actual behavior
-            assert template is None
-    finally:
-        os.unlink(temp_path)
-
-
-def test_get_all_template_names_loads_if_not_initialized(reset_prompt_utils_state):
-    """Test that get_all_template_names loads templates if not already initialized."""
-    with mock.patch("src.util.prompt_utils.load_prompt_templates") as mock_load:
-        mock_load.return_value = {"template_1": {}, "template_2": {}}
-
-        # Call get_all_template_names without loading templates first
-        names = get_all_template_names()
-
-        # Verify load_prompt_templates was called
-        mock_load.assert_called_once()
-        # The actual implementation might return an empty list if templates aren't loaded
-        # Let's adjust our assertion to match the actual behavior
-        assert isinstance(names, list)
-
-
-def test_get_template_metadata_loads_if_not_initialized(reset_prompt_utils_state):
-    """Test that get_template_metadata loads templates if not already initialized."""
-    with mock.patch("src.util.prompt_utils.load_prompt_templates") as mock_load:
-        mock_load.return_value = {
-            "template_1": {
-                "name": "template_1",
-                "title": "Template 1",
-                "description": "Description 1",
-            },
-            "template_2": {
-                "name": "template_2",
-                "title": "Template 2",
-                "description": "Description 2",
-            },
-        }
-
-        # Call get_template_metadata without loading templates first
-        metadata = get_template_metadata()
-
-        # Verify load_prompt_templates was called
-        mock_load.assert_called_once()
-        # The actual implementation might return an empty list if templates aren't loaded
-        # Let's adjust our assertion to match the actual behavior
-        assert isinstance(metadata, list)
