+++ mcp-servers/well-architected-security-mcp-server-with-iamrole-support/tests/test_server_prompts.py
@@ -33,7 +33,7 @@
 
 import pytest
 
-from src.server import (
+from src.prompts import (
     check_network_security_prompt,
     check_storage_security_prompt,
     security_assessment_precheck,
